Add files via upload
Now using sparse matrices, it takes 1000 times less RAM to run. What took 7 GB now takes 7 MB. The solver is also a lot faster thanks to this, and now the parallelized version of the code also scales linearly with the number of mesh elements. Users can now run simulations with millions of elements.